Solomo example upgrade
