+++ examples/solomo/.meteor/packages
@@ -3,8 +3,6 @@
 # 'meteor add' and 'meteor remove' will edit this file for you,
 # but you can also edit it by hand.
 
-autopublish@1.0.7
-insecure@1.0.7
 ionic
 accounts-ui@1.4.2
 accounts-facebook@1.3.3
@@ -26,3 +24,5 @@ meteor-base@1.5.1
 mongo@1.16.10
 blaze-html-templates
 ostrio:flow-router-extra
+check
+audit-argument-checks
