+++ examples/solomo/.meteor/versions
@@ -4,7 +4,7 @@ accounts-oauth@1.4.4
 accounts-ui@1.4.2
 accounts-ui-unstyled@1.7.1
 allow-deny@1.1.1
-autopublish@1.0.7
+audit-argument-checks@1.0.7
 autoupdate@1.8.0
 babel-compiler@7.10.5
 babel-runtime@1.5.1
@@ -40,7 +40,6 @@ hot-code-push@1.0.4
 html-tools@1.1.4
 htmljs@1.2.1
 id-map@1.1.1
-insecure@1.0.7
 inter-process-messaging@0.1.1
 ionic@1.0.0
 jquery@3.0.0
