+++ examples/solomo/server/main.js
@@ -1,0 +1,8 @@
+import { Meteor } from 'meteor/meteor';
+import { Photos } from '../devshop-demo'
+import '../devshop-demo';
+
+Meteor.publish('photos', function () {
+  return Photos.find({}, {sort: {"createdAt": -1}});
+})
+
